Fixed some differences between XML output (as displayed by sadf -x) and DTD document.
In the DTD document, cpu-frequency is defined as a list of cpufreq items.
But sadf -x outputs:

<cpu-frequency unit="MHz">
	<cpu number="all" frequency="3000.00"/>
	<cpu number="0" frequency="3000.00"/>
	<cpu number="1" frequency="3000.00"/>
</cpu-frequency>

So replace cpu by cpufreq in sadf.c (we don't keep the name
"cpu" as it is already used in CPU statistics):

<cpu-frequency unit="MHz">
	<cpufreq number="all" frequency="3000.00"/>
	<cpufreq number="0" frequency="3000.00"/>
	<cpufreq number="1" frequency="3000.00"/>
</cpu-frequency>

In the DTD document, fan-speed is defined as a list of fanspd items.
But sadf -x outputs:

<fan-speed unit="rpm">
	<fan number="1" rpm="1869" drpm="1869" device="dme1737-i2c-1-2e"/>
	<fan number="2" rpm="1136" drpm="1136" device="dme1737-i2c-1-2e"/>
	<fan number="3" rpm="0" drpm="0" device="dme1737-i2c-1-2e"/>
</fan-speed>

The name "word" can be kept here, so replace fanspd by fan in DTD document.

Note that XSD document is correct.