No longer assume that device-mapper major number is 253. Get the real number from /proc/devices file.
The sar, sadf and iostat commands used to assume that device-mapper
major number was 253. This happened to be false sometimes. So get
the real number from the /proc/devices file.

From Mike Coleman <tutufan@gmail.com> 04/10/2010:

The iostat program seems to assume that the major device number for
devmap will always be 253.  This doesn't seem to be an official
number, though, and I have a box where it actually ends up being 252,
which breaks 'iostat -N'.

It looks like this can be determined dynamically by looking at /proc/devices.