Option "-P ON" added to mpstat.
This option tells mpstat to display statistics for online
processors only.
mpstat manual page updated.

On 06/30/2011 06:41 AM, Ananth N Mavinakayanahalli wrote:

a. Consistency with output of top and /proc/cpuinfo, both of which won't
     display information of offlined CPUs.
b. On POWER7 for instance, with SMT, when SMT is turned off, 3 of the 4
     CPUs are off-lined. Someone using just mpstat will have no idea that
     he has just 1 CPU underneath while mpstat says he has 4, without an
     indication of which ones are usable and which ones are not.

I think, at the least, the -P switch should be educated with a new
option that displays only online CPU information, if there is a hard
requirement to preserve existing functionality.