XSD document updated.
A maxOccurs indicator has been added for the timestamp element.
This is indicator is set to "unbounded", which is compulsory here
as the default value is 1 if not specified.

Mail from Frank Glinka <glinkaf@uni-muenster.de> 24/04/2012:

There seem to be gaps/inconsistencies between the produced XML and the  sysstat.xsd that you provide on your website. The xsd does not specify any maximum occurrence indicators, which in that case default to '1'. For example, although the 'timestamp' element occurs multiple times within the 'statistics' element inside the XML, the xsd claims it appears exactly once. In that case, a 'maxOccurs="unbounded"' is required. (This is an issue if I use Java's JAXB and your sysstat.xsd to read & parse the XML file automatically.)

See:
http://www.w3.org/TR/2004/REC-xmlschema-1-20041028/structures.html#Particle_details
http://www.w3schools.com/schema/schema_complex_indicators.asp