Change time format from HH-MM-SS to HH:MM:SS in the reports displyaed by sadf.
Change the time format to HH:MM:SS to be consistent with XSD
document. This was already the format used a few versions ago
and was changed for an unknown reason.

Mail from Frank Glinka <glinkaf@uni-muenster.de> 24/04/2012:

Your sysstat.xsd specifies the type xs:time for the time elements within the XML. According to the schema specification the format then must be HH:MM:SS for the data but sadf's format is HH-MM-SS. Correspondingly, the element is not valid and not parsed by JAXB.

See:
http://www.w3.org/TR/xmlschema-2/#time
http://www.w3schools.com/schema/schema_dtypes_date.asp