Rename nfsiostat command to nfsiostat-sysstat
nfsiostat was added to the sysstat package in 2010, but such a command
has already existed in the nfs-utils package since 2008.
So to avoid confusion, rename nfsiostat to nfsiostat-sysstat and
indicate it is now obsolete.
The nfsiostat command from the sysstat package will be removed in a
future version since we don't need two similar commands.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>