Change file_magic structure contents
Move header_size field to the end of the structure, so that other
programs can still tell it is a sysstat sar data file, even though
the format has changed.

Also, to avoid causing format changes, add padding to the file_magic
structure so that when they grow in size other tools can still
handle the existing data.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>