Rename --disable-man-group option to --disable-file-attr
This option was originally intended to prevent the configuration process
from setting the group ID for manual pages that were installed.
It is now also used to tell the build process to not set the user and
group of the /var/log/sa directory when it is created.
This option may be useful when creating an RPM package with a non
privileged user.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>