Refactor systemd unit files to better work with suspend/resume
Previously a comment was written to the accounting log only on system boot,
which caused artifacts in generated graphs of laptops with suspend/resume
functionality. The solution is to write a similar comment both before suspend
and after resume, so that the graphing program can insert a gap in that
timeframe.

As a nice side-effect, one can now start sysstat just by doing
"systemctl start sysstat", and the timer unit will be pulled automatically
in the proper order, so that there is no contention for locking
the accounting file. Same goes for stopping.