sadf: Fix bug preventing sadf from converting an old format datafile
Version 11.1.5 added a check on file_magic.header_size present in the
header of a data file.
An error message should not be displayed if this check is not successful
when sadf is used to convert the data file (sadf -c) or used to display
its header contents (sadf -H).

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>