JSON: Don't display fchosts tag when no stats exist
The JSON print function can be called for A_NET_FC (Fibre Channel)
activity even if no statistics exist. This is because this activity has
the AO_CLOSE_MARKUP flag used to close the "network" section of the
output. Anyway, in this case (when no stats exist), it should only do
that (close the network section) et not display the "fchost" tag.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>