SVG: Add SVG output for task creation and system switching activity.
Update sadf's SVG ouptput to create graphs for task creation and system
switching activity (this corresponds to the output of sar -w).
This is just a proof of concept.
Try it with "sadf -g /your/sa/datafile -- -w > out.html" and open
the resulting HTML file in your web browser.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>