SVG: Handle dynamically (un)registered network interfaces
Network interfaces may be dynamically registered or unregistered.
That is to say they can appear in the /proc/net/dev file when
the corresponding device is connected, and also disappear when
the device is removed.

This patch tries to handle this so that graphs are drawn for
the interfaces which are connected on the fly. Graphs are drawn only
when the device is actually connected, which means that they may be
discontinuous if the device is connected, then disconnected, then
reconnected again, just like when a system restart occurrs.

At the present time, this patch assumes that the devices are always at
the same position in /proc/net/dev, which is not necessarily the case.
This will be fixed later.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>