sar: Fix possible compatibility issues between sysstat versions
Fix an issue that made sar unable to read statistics for network
and disk devices from a datafile created by a more recent version of
sysstat when the size of the structures used to save those statistics
was greater than that of the current version.
This may be the case when eg. new fields are appended to network or disk
structures.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>