SVG: Make SVG code locale independent
Graphs could not be displayed if corresponding SVG code had been
generated on machines with specific locales.

Eg: If you set your locale to, say, fr_FR then here is a sample SVG
code you could get:

...
<path id="g9p0" d=" M311,2663,54 L613,1446,20 L913,423,75 (...)
...

To avoid the confusion, this patch sets the LC_NUMERIC to "C" so that a
decimal point is used instead of a comma:

...
<path id="g9p0" d=" M311,2663.54 L613,1446.20 L913,423.75 (...)
...

Resulting SVG output can now be displayed whatever locale is used.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>