SVG: Set a min value on Y axis
Set a min value on Y axis so that graduations remain meaningful.
For instance, if the max value for a graph was 0.01 then the graduations
on the Y axis were 0.01, 0.01 and 0.01 (for 3 horizontal lines). In this
case, this patch will set the max value to 0.03 to get some meaningful
graduations values: 0.01, 0.02 and 0.03.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>