SVG: Add "autoscale" option
This new option (to be used with sadf's switch -O) tells sadf to draw
all the graphs of a given view as large as possible based on current
view's scale. To do this, a factor (10, 100, 1000...) is used to
enlarge the graph drawing.
This option may be intesting when several graphs are drawn on the same
view, some with only very small values, and others with high ones,
the latter making the former difficult to see.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>