sar: Allow selection of a range of CPU
Update sar so that the user can now select a range of CPUs with option -P
(in addition to selecting an individual CPU). It is also possible to
select CPU "all" which is the global average among all CPUs (not to be
confused with the "ALL" keyword which selects all the CPUs of the
machine).

E.g.:

sar -P 0,3-5,12-	Selects CPUs #0, #3, #4, #5, and all the CPUs
from #12 and following.

sar -P all,1		Select CPU #1 and the global average CPU value
among all processors.

This patch will be reworked to cover mpstat.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>