sar: Better handle dynamically registered devices
When a newly registered device (network interface or disk) was
supernumerary (i.e. there was no slot left in the list to store it), it
would take the place of another device in the list. This other device
would also appear as new and its counters would jump as if starting from
zero. This patch fixes that.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>