SVG: Give actual canvas height
Add a comment at the end of SVG output that indicates the actual canvas
height.
The actual height is frequently different (i.e. smaller) than that
indicated in the header of the SVG output. This is because some
activities have pre-allocated entries that don't result in graphs being
drawn, or because option 'skipempty' has been used.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>