Close #138: SVG: Allow multiple charts on a row
Add a new option ("packed") to tell sadf to put multiple charts on a row
when generating SVG graphs.
In fact, with this option, all charts from the same activity (and for
the same device) will appear on the same row.

E.g.: sadf -g -O packed sa01 -- -A > sa01.svg

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>