fix 00 byte after values when --human is not set
Since 33557c3d: Display a percent sign after the value when --human
option used in non --human mode the output has an exta spurious '0'
byte.

That is printed by the percentage unit being set to char u = '\0'.
But that is not "nothing" it is literally byte "00".

This change only calls printf with a literal % when needed.

Fixes 165

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>