sar/sadf: On-the-fly endianness conversion
This patch aims at making sar and sadf able to read and process files
created on machines with a different endian type. It will now be
possible e.g., to read a file with sar on a little-endian machine even
if this has been created on a big-endian machine.
Yet it will still not be allowed to append (write) data with sar to a
file with a different endianness type.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>