SVG: Fix segfault triggered by a varying number of CPU in file
Buffer allocation must be based on the maximum number of CPU found in
file. "A_CPU" being a volatile activity, its number is calculated each
time a LINUX RESTART is inserted in file and it may happen that this
number is higher than the first result we got when the file was created.

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>