sar: Test for zero value when reallocating all the buffers
When sar reads the contents of a file and meets a LINUX RESTART, it
may have to reallocate the buffers used to save CPU statistics to match
the new number of CPU. If the number of CPU has increased it doubles the
size of its buffers.
There is a problem though if CPU activity was not included in file (or
if its format was unknown to current sysstat version). In this case the
buffer size was zero. So test this before reallocating buffers.

Below is a sample output before the patch was applied.
The file contains only A_PCSW activity then a LINUX RESTART message
(which cannot be displayed):

$ sar -f data0 -w

Linux 4.4.14-200.fc22.x86_64 (test.home) 	03/04/18 	_x86_64_
(8 CPU)

09:31:50       proc/s   cswch/s
09:31:51         0.00    402.00
09:31:52         0.00    356.00
Average:         0.00    379.00
srealloc	<------------------------- Error!

Signed-off-by: Sebastien GODARD <sysstat@users.noreply.github.com>